test(robot): fix default backup target is not available error
