+++ e2e/tests/negative/uninstallation_checks.robot
@@ -60,7 +60,7 @@ Uninstallation Checks
 
     # Assume this is another Longhorn cluster
     Then Install Longhorn
-    And set_backupstore
+    And Set default backupstore
     And Check backup synced from backupstore
     When Create DR volume 0 from backup 0 in another cluster
     And Wait for volume 0 restoration from backup 0 in another cluster completed
